+++ app/models/alaveteli_pro/embargo_extension.rb
@@ -14,7 +14,6 @@ module AlaveteliPro
   class EmbargoExtension < ApplicationRecord
     belongs_to :embargo,
                inverse_of: :embargo_extensions
-    validates_presence_of :embargo
     validates_presence_of :extension_duration
     validates_inclusion_of :extension_duration,
                            in: lambda { |_e| AlaveteliPro::Embargo.new.allowed_durations }