+++ app/models/info_request.rb
@@ -67,15 +67,18 @@ class InfoRequest < ApplicationRecord
 
   belongs_to :user,
              inverse_of: :info_requests,
-             counter_cache: true
+             counter_cache: true,
+             optional: true
 
   validate :must_be_internal_or_external
 
   belongs_to :public_body,
              inverse_of: :info_requests,
-             counter_cache: true
+             counter_cache: true,
+             validate: false
   belongs_to :info_request_batch,
-             inverse_of: :info_requests
+             inverse_of: :info_requests,
+             optional: true
 
   validates_presence_of :public_body, message: N_("Please select an authority")