+++ app/models/notification.rb
@@ -24,7 +24,7 @@ class Notification < ApplicationRecord
   DAILY = :daily
   enum frequency: [ INSTANTLY, DAILY ]
 
-  validates_presence_of :info_request_event, :user, :frequency, :send_after
+  validates_presence_of :frequency, :send_after
 
   before_validation :calculate_send_after