+++ app/models/outgoing_message.rb
@@ -42,7 +42,6 @@ class OutgoingMessage < ApplicationRecord
   attr_accessor :default_letter
 
   before_validation :cache_from_name
-  validates_presence_of :info_request
   validates_presence_of :from_name, unless: -> (m) { !m.info_request&.user }
   validates_inclusion_of :status, in: STATUS_TYPES
   validates_inclusion_of :message_type, in: MESSAGE_TYPES
@@ -56,7 +55,8 @@ class OutgoingMessage < ApplicationRecord
   belongs_to :incoming_message_followup,
              inverse_of: :outgoing_message_followups,
              foreign_key: 'incoming_message_followup_id',
-             class_name: 'IncomingMessage'
+             class_name: 'IncomingMessage',
+             optional: true
 
   has_one :user,
           inverse_of: :outgoing_messages,